Getting it formatted how I want then I'll deal with the mapping error
